Check if the `window.performance` should be used in a duck-typy way
In my Rollup setup, there is a fake `global` object that is a reference to `window`, which means that the polyfill (meant for Node.js) will be used instead of `window.performance`.
I'd rather check if `window` exists and has a `performance()` method, and then only use the polyfill if it's not the case.